Added classes to work with fractions, elements in the field Q[sqrt(d)], and elements in the ring Z[sqrt(d)]. Added a common functions class and solutions to problems 61-78

+++ src/org/euler/main/Problem60.java
@@ -197,7 +197,7 @@ public class Problem60 {
 		}
 		//----------------
 		long deltaTime = System.nanoTime() - lastTime;	
-		System.out.println("Computed in " + ((deltaTime/1000000000)/60) + " minutes.");
+		System.out.println("Computed in " + ((double)(deltaTime/1000000000)/60.0) + " minutes.");
 	}
 
 }
